+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog of threedigrid-builder
 0.8.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Only process cross section definitions that are actually used.
+
+- Removed cross1 & cross2 in the in-memory or geopackage output and added cross_id1 and
+  cross_id2.
 
 
 0.8.2 (2021-12-05)