+++ CHANGES.rst
@@ -1,7 +1,26 @@
 Changelog of threedigrid-builder
 ================================
 
-0.8.1 (unreleased)
+0.8.3 (unreleased)
+------------------
+
+- Only process cross section definitions that are actually used.
+
+- Removed cross1 & cross2 in the in-memory or geopackage output and added cross_id1 and
+  cross_id2.
+
+- Fixed bug where writing a single line geometry or geometries of equal size would
+  result in an incorrect cast to a numpy object dtype.
+
+
+0.8.2 (2021-12-05)
+------------------
+
+- Fixed the Linux wheel distribution. These are now built with manylinux2014 instead of
+  manylinux2010.
+
+
+0.8.1 (2021-12-05)
 ------------------
 
 - Added support for zero-d administration including surfaces and impervious surfaces.